+++ R/handler_progress.R
@@ -65,7 +65,8 @@ handler_progress <- function(format = ":spin [:bar] :percent :message", show_aft
       new = function(...) list(
         finished = FALSE,
         tick = function(...) NULL,
-        update = function(...) NULL
+        update = function(...) NULL,
+        message = function(...) NULL
       )
     )
     get_private <- function(pb) NULL
@@ -148,8 +149,9 @@ handler_progress <- function(format = ":spin [:bar] :percent :message", show_aft
         if (!state$enabled || config$times <= 2L) return()
         make_pb(format = format, total = config$max_steps,
                 clear = config$clear, show_after = config$enable_after)
-        if (inherits(progression, "sticky") && length(state$message) != 0)
+        if (inherits(progression, "sticky") && length(state$message) > 0) {
           pb$message(state$message)
+        }
         pb_tick(pb, state$delta, message = state$message)
       },