+++ R/options.R
@@ -29,7 +29,7 @@
 #'   \item{\option{progressr.enable}:}{
 #'     (logical)
 #'     If FALSE, then progress is not reported.
-#'     (Default: TRUE)
+#'     (Default: TRUE in interactive mode, otherwise FALSE)
 #'   }
 #'
 #'   \item{\option{progressr.enable_after}:}{