+++ R/slow_sum.R
@@ -9,6 +9,9 @@
 #'
 #' @param message If TRUE, then a message is outputted per element.
 #'
+#' @param sticky If TRUE, then a "sticky" message is outputted every
+#' ten element.
+#'
 #' @return The sum of all elements in `x`.
 #'
 #' @section Progress updates:
@@ -16,7 +19,7 @@
 #'
 #' @keywords internal
 #' @export
-slow_sum <- function(x, delay = getOption("progressr.demo.delay", 1.0), stdout = FALSE, message = TRUE) {
+slow_sum <- function(x, delay = getOption("progressr.demo.delay", 1.0), stdout = FALSE, message = TRUE, sticky = TRUE) {
   p <- progressor(along = x)
 
   sum <- 0
@@ -24,7 +27,6 @@ slow_sum <- function(x, delay = getOption("progressr.demo.delay", 1.0), stdout =
     p(amount = 0)   ## "I'm alive" progression update
     Sys.sleep(0.2*delay)
     if (stdout) cat(sprintf("O: Element #%d\n", kk))
-    p(amount = 0)
     Sys.sleep(0.2*delay)
     p(amount = 0)
     Sys.sleep(0.2*delay)
@@ -34,6 +36,13 @@ slow_sum <- function(x, delay = getOption("progressr.demo.delay", 1.0), stdout =
     if (message) message(sprintf("M: Added value %g", x[kk]))
     p(amount = 0)
     Sys.sleep(0.2*delay)
+    if (sticky && kk %% 10 == 0) {
+      p(
+        amount = 0,
+        message = sprintf("P: %d elements done", kk),
+        class = "sticky"
+      )
+    }
   }
 
   p(amount = 0)