+++ R/with_progress.R
@@ -36,9 +36,9 @@
 #' @example incl/with_progress.R
 #'
 #' @details
-#' _IMPORTANT: This function is meant for end users only.  It should not
+#' **IMPORTANT: This function is meant for end users only.  It should not
 #' be used by R packages, which only task is to _signal_ progress updates,
-#' not to decide if, when, and how progress should be reported._
+#' not to decide if, when, and how progress should be reported.**
 #'
 #' @section Progression handler functions:
 #' Formally, progression handlers are calling handlers that are called